app/requirements.txt
Flask==3.0.3
pandas==2.2.2
pip==22.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
